Update to use 3550-3600 MHz.
This required understanding how SSB locations are chosen in NR. (I
couldn't get the COTS UEs in the lab to attach using the same gNB
settings that worked for the SDR-based UE. Always read the spec!)

Since the number of ARFCNs available is quite large in NR, the spec
allows only particular frequencies for SSB locations to limit the search
space for UEs.

Since we are using band n78, the range of interest in the spec is
3000MHz - 24250MHz, for which the allowed SSB locations are:

3000MHz + N * 1.44MHz (from 3GPP TS 38.101-1 v15.17.0 Table 5.4.3.1-1)

where N is an integer in the range 0 - 14756.

We let N be 400, which corresponds to an SSB location of 3576MHz (ARFCN
638400). We set dl_absoluteFrequencyPointA (defines lowest occupied PRB)
to 3556.92MHz (ARFCN 637128) and leave the bandwidth settings at 106
PRBs (40MHz) to make sure we are only transmitting in the 3550MHz -
3600MHz range. This works just fine with the Quectel RM500QGL modules.